backend/requirements.txt
fastapi
uvicorn
numpy
tensorflow
python-multipart
Pillow
scikit-learn